requirements.txt
numpy==1.20.3
PyGObject==3.36.0
Pillow==7.0.0